Start adding image build
Multistage as it should be. Rather interesting hoops to jump through in order to get deps to download and build in their own step: a dummy sourcefile is needed, _and_ the build of the main package needs to be explicitly cleaned afterwards, or the old build artefacts will stick around and the dummy will get rebuilt.